scrapers/requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
attrs==18.2.0
Automat==0.7.0
cffi==1.11.5
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.5
mccabe==0.6.1
parsel==1.5.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==2.1.1
pyOpenSSL==18.0.0
python-dateutil==2.8.0
queuelib==1.5.0
Scrapy==1.5.1
service-identity==17.0.0
six==1.11.0
Twisted==19.2.0
typed-ast==1.1.0
w3lib==1.19.0
wrapt==1.10.11
zope.interface==4.5.0
